make bet

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->string('name')->unique();
             $table->string('email')->unique();
             $table->integer('level')->default(0);
-            $table->integer('coins')->default(0);
+            $table->double('coins')->default(0);
             $table->integer('noloses')->default(0);
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
